Using a child context with private queue concurrency type.
Quoting from https://developer.apple.com/library/mac/documentation/Cocoa/Conceptual/CoreData/Concurrency.html
"In general, avoid doing data processing on the main queue that is not user-related. [...] create a private queue context and perform the import on the private context".

To make this easier and more organized, there is now a CoreDataStack class that
takes care of instantiating all the Core Data stuff and provides a method for
creating private queue contexts that are children from the main queue context.

The slow method is now executed in the private queue, thus it does not block
the main thread, neither does it crash due to multithreading violation :)