Added table view editing mode to delete objects (very slow method, again!)
The table view is getting more functionality, so now the slow insertion method
was implemented as a pull-to-refresh and in its place, the left navigation bar
item is an Edit button that allows items to be deleted.

However, this deletion method is slow for some reason (i.e. the server needs to
authorize it) so, once again, it should not be done in the main thread. A less
naive approach was done this time: a child context was created and the
deletion is done from inside it. All good? No.

Terminating app due to uncaught exception 'NSInvalidArgumentException', reason: 'An NSManagedObjectContext cannot delete objects in other contexts.'