Accessing object from another context through its objectID.
objectID is thread-safe, so it can be used to access an object from a MOC in
another MOC. Problem solved: no hangs, no crashes.

However... What a bad UX to make the user keep waiting until the item disappear!
Unfortunately, we can't delete the object if the server has not yet authorized
it. What can we do, then? If only we could delete it from the UI, but keep it
in the database somehow...

Well, turns out we can - next commit!