The MOC connected to the PSC is not in the main thread anymore.
We freed the main queue MOC from dealing with slow methods, but this MOC was
still connected to the PSC, which requires disk access. So now we have a new
MOC specifically for this, the master MOC, and the main queue MOC is its child.

Now the saveRecursively method is really recursive, as there are more than 2
MOCs in the stack (master -> main -> private).

We have a problem though. When we try to delete an item:
*** Terminating app due to uncaught exception 'NSInternalInconsistencyException', reason: 'Error while trying to access object to delete in the private MOC

The assertion that was put right after existingObjectWithID:error: failed. This
is because the objectID obtained is a temporary one. This can be directly seen
by its format, e.g.:
<x-coredata:///APLEvent/t9D2EB09F-817D-431A-B14F-75FCFFF4E9B52>

If it was a persistent objectID, it would start with a guid (identifying the
PSC), followed by the entity's name (APLEvent).