Using a scratchpad context for the UI and rolling back changes when necessary.
- Added to CoreDataStack a method for creating main queue contexts children from
the default main queue context: createScratchpadMoc
- Added to CoreDataStack a method for creating private contexts that are
children from a specified context. Mostly useful when you are using a scratchpad
- The table view is now linked to a scratchpad context, which is instantly
updated when the user requests a deletion, but does not save the change to disk
until the deletion is authorized by some slow processing. The slow processing is
moved to be done after the UI modifications.
- If the deletion is not authorized, the change is rolled back via -rollback. Is
that correct? Should we really revert ALL the unsaved changes? What if the user
quickly requests to delete several objects and only some of them were denied?