Added an UndoManager to the uiMoc. Now undo works.
However, the behavior is odd when the user quickly requests to delete several
objects and only some of them are denied. The wrong changes are being reversed.
Of course: we are reverting the "last" unsaved changes, but other deletions have
happened since the last one, so "the last change" is not the correct way to
refer to it.