Adding test to demonstrate the importance of avoiding massive saves.
My results for the tests were:
1) Massive insertions and saves (one save for each of the 1000 insertions):
1.453 sec (21% STDEV)
2) Massive insertions with single save: 0.080 sec (31% STDEV)

So, executing save in the very end is most usually the best practice.

Nonetheless, there is still something that we are doing more often than needed:
saving to disk! The saveRecursively method is saving to the PSC every time. Can
this be reduced? What are the performance impacts?