Adding a dummy disk write reducer: saves are only done 1/10th of the times.
Such a dummy modification is not SO dangerous. From the previous commits, it was seen
that iOS automatically saves to the PSC when the app is closed, even if force closed
(by swiping up in the multitasking screen). *However if the app crashes, iOS will NOT
save to disk!!!*
So, this dummy modification is simply not suitable for production apps. Even your
"perfect" app WILL crash. Ok ok, maybe not, but you should always consider that.

Anyway, the point here is to show what's the impact on performance on saving to disk.
My new results for the tests were:
1) Massive insertions and saves (one save for each of the 1000 insertions):
1.453 sec (21% STDEV) - 74% decrease.
2) Massive insertions with single save: 0.037 sec (91% STDEV) - 54% decrease